Improve documentation throughout all packages.
