+++ manager/src/main/java/io/atomix/manager/util/ResourceManagerTypeResolver.java
@@ -13,12 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License
  */
-package io.atomix.manager;
+package io.atomix.manager.util;
 
 import io.atomix.catalyst.serializer.SerializableTypeResolver;
 import io.atomix.catalyst.serializer.SerializerRegistry;
 import io.atomix.manager.state.*;
-import io.atomix.resource.*;
+import io.atomix.resource.ResourceException;
+import io.atomix.resource.ResourceType;
+import io.atomix.resource.util.InstanceTypeResolver;
+import io.atomix.resource.util.ResourceCommand;
+import io.atomix.resource.util.ResourceQuery;
+import io.atomix.resource.util.ResourceRegistry;
 
 /**
  * Resource manager serializable type resolver.
@@ -40,8 +45,8 @@ public class ResourceManagerTypeResolver implements SerializableTypeResolver {
     // Register resource state machine types.
     registry.register(ResourceCommand.class, -50);
     registry.register(ResourceQuery.class, -51);
-    registry.register(ResourceStateMachine.ConfigureCommand.class, -52);
-    registry.register(ResourceStateMachine.DeleteCommand.class, -53);
+    registry.register(ResourceCommand.Configure.class, -52);
+    registry.register(ResourceCommand.Delete.class, -53);
 
     // Register resource types.
     for (ResourceType type : resourceRegistry.types()) {
