+++ manager/src/main/java/io/atomix/resource/util/ResourceRegistry.java
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License
  */
-package io.atomix.resource;
+package io.atomix.resource.util;
 
 import io.atomix.catalyst.util.Assert;
+import io.atomix.resource.ResourceType;
 
 import java.util.Collection;
 import java.util.Map;
