+++ manager/src/main/java/io/atomix/resource/util/ServiceLoaderResourceResolver.java
@@ -13,9 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License
  */
-package io.atomix.resource;
+package io.atomix.resource.util;
 
 import io.atomix.catalyst.CatalystException;
+import io.atomix.resource.Resource;
+import io.atomix.resource.ResourceType;
 
 import java.io.BufferedReader;
 import java.io.IOException;
