+++ pom.xml
@@ -22,8 +22,8 @@
     <java.version>1.8</java.version>
     <slf4j.version>1.7.7</slf4j.version>
     <logback.version>1.1.2</logback.version>
-    <catalyst.version>1.0.1</catalyst.version>
-    <copycat.version>1.0.0-rc2</copycat.version>
+    <catalyst.version>1.0.2-SNAPSHOT</catalyst.version>
+    <copycat.version>1.0.0-SNAPSHOT</copycat.version>
 
     <maven.source.plugin.version>2.2.1</maven.source.plugin.version>
     <maven.compiler.plugin.version>3.0</maven.compiler.plugin.version>
@@ -149,7 +149,7 @@
           <doctitle>Atomix API Reference (${jv})</doctitle>
           <windowtitle>Atomix API Reference (${jv})</windowtitle>
           <aggregate>true</aggregate>
-          <excludePackageNames>io.atomix.examples*:*.state*:io.atomix.manager*:io.atomix.testing*</excludePackageNames>
+          <excludePackageNames>io.atomix.examples*:*.state*:*.util*:io.atomix.manager*:io.atomix.testing*</excludePackageNames>
           <additionalparam>
             -Xdoclint:none
             -notimestamp
