+++ resource/src/main/java/io/atomix/resource/ResourceTypeInfo.java
@@ -23,7 +23,13 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Resource type info.
+ * Annotation for specifying resource state and serialization information.
+ * <p>
+ * {@link Resource}s <em>must</em> be annotated with this annotation to be managed by the Atomix cluster.
+ * This annotation provides information about the resource's {@link ResourceStateMachine state machine}
+ * and {@link SerializableTypeResolver serialization} of resource commands. Atomix will use the annotation
+ * at runtime to create a {@link ResourceType} with which it can identify the resource type and create
+ * server-side replicated state machines and register serializable types relevant to the resource.
  *
  * @author <a href="http://github.com/kuujo>Jordan Halterman</a>
  */
@@ -32,17 +38,32 @@ import java.lang.annotation.Target;
 public @interface ResourceTypeInfo {
 
   /**
-   * The resource type ID.
+   * The 8-bit resource type ID.
+   * <p>
+   * This identifier must be unique across all resource types and should be within the range
+   * {@code -128} to {@code 127}. Core Atomix resource types are typically identified with
+   * negative numbers, so the range {@code 1} through {@code 127} is reserved for user-provided
+   * resource types.
    */
   int id();
 
   /**
    * The resource state machine class.
+   * <p>
+   * This is the state machine class that Atomix will construct on each server in the cluster.
+   * The state machine is responsible for managing state on each server. Each resource instance
+   * has an associated state machine which will last throughout its lifetime.
    */
   Class<? extends ResourceStateMachine> stateMachine();
 
   /**
    * The resource serializable type resolver.
+   * <p>
+   * Resource types should provided a {@link SerializableTypeResolver} to specify how to serialize
+   * {@link io.atomix.copycat.Command commands} and {@link io.atomix.copycat.Query queries} and any
+   * other types of objects required to operate the resource. Atomix will query the type resolver
+   * to ensure serializable types are registered on both clients and servers prior to usage of the
+   * resource.
    */
   Class<? extends SerializableTypeResolver> typeResolver();
